+++ spring-boot-mvc/pom.xml
@@ -97,12 +97,6 @@
             <version>${spring.fox.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.apache.tomcat.embed</groupId>
-            <artifactId>tomcat-embed-jasper</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
         <!-- AOP -->
         <dependency>
             <groupId>org.aspectj</groupId>
