+++ spring-boot-mvc/src/test/java/org/baeldung/SpringContextLiveTest.java
@@ -6,7 +6,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-import com.baeldung.autoconfiguration.MySQLAutoconfiguration;
+import com.baeldung.springbootannotations.MySQLAutoconfiguration;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = MySQLAutoconfiguration.class)
